Add JWT token validation to UserController

+++ src/main/java/dev/hiruna/edutrack/controller/UserController.java
@@ -2,7 +2,10 @@ package dev.hiruna.edutrack.controller;
 
 import dev.hiruna.edutrack.dto.UserDTO;
 import dev.hiruna.edutrack.service.UserService;
+import dev.hiruna.edutrack.util.JWTAuthenticator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,28 +17,50 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private JWTAuthenticator jwtAuthenticator;
+
     @GetMapping
-    public List<UserDTO> getAllUsers() {
-        return userService.getAllUsers();
+    public ResponseEntity<List<UserDTO>> getAllUsers(@RequestHeader("Authorization") String authHeader) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            return new ResponseEntity<>(userService.getAllUsers(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/{id}")
-    public UserDTO getUserById(@PathVariable Integer id) {
-        return userService.getUserById(id);
+    public ResponseEntity<UserDTO> getUserById(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            UserDTO userDTO = userService.getUserById(id);
+            return new ResponseEntity<>(userDTO, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PostMapping
-    public UserDTO createUser(@RequestBody UserDTO userDTO) {
-        return userService.createUser(userDTO);
+    public ResponseEntity<UserDTO> createUser(@RequestHeader("Authorization") String authHeader, @RequestBody UserDTO userDTO) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            UserDTO createdUser = userService.createUser(userDTO);
+            return new ResponseEntity<>(createdUser, HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping("/{id}")
-    public UserDTO updateUser(@PathVariable Integer id, @RequestBody UserDTO userDTO) {
-        return userService.updateUser(id, userDTO);
+    public ResponseEntity<UserDTO> updateUser(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id, @RequestBody UserDTO userDTO) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            UserDTO updatedUser = userService.updateUser(id, userDTO);
+            return new ResponseEntity<>(updatedUser, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Integer id) {
-        userService.deleteUser(id);
+    public ResponseEntity<Void> deleteUser(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            userService.deleteUser(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
